+++ knf/united/united_test.go
@@ -80,7 +80,7 @@ func (s *UnitedSuite) TestKNFOnly(c *C) {
 	c.Assert(GetTZ("test:timezone").String(), Equals, "Europe/Zurich")
 	c.Assert(GetL("test:list"), DeepEquals, []string{"Test1", "Test2"})
 
-	err := Combine(
+	Combine(
 		Mapping{"test:string", "test-string", "TEST_STRING"},
 		Mapping{"test:integer", "test-integer", "TEST_INTEGER"},
 		Mapping{"test:float", "test-float", "TEST_FLOAT"},
@@ -92,8 +92,6 @@ func (s *UnitedSuite) TestKNFOnly(c *C) {
 		Mapping{"test:timezone", "test-timezone", "TEST_TIMEZONE"},
 		Mapping{"test:list", "test-list", "TEST_LIST"},
 	)
-
-	c.Assert(err, IsNil)
 
 	c.Assert(GetS("test:string"), Equals, "Test")
 	c.Assert(GetI("test:integer"), Equals, 123)
@@ -143,7 +141,7 @@ func (s *UnitedSuite) TestWithOptions(c *C) {
 
 	c.Assert(errs, HasLen, 0)
 
-	err := Combine(
+	Combine(
 		Mapping{"test:string", "test-string", "TEST_STRING"},
 		Mapping{"test:integer", "test-integer", "TEST_INTEGER"},
 		Mapping{"test:float", "test-float", "TEST_FLOAT"},
@@ -155,8 +153,6 @@ func (s *UnitedSuite) TestWithOptions(c *C) {
 		Mapping{"test:timezone", "test-timezone", "TEST_TIMEZONE"},
 		Mapping{"test:list", "test-list", "TEST_LIST"},
 	)
-
-	c.Assert(err, IsNil)
 
 	optionHasFunc = opts.Has
 	optionGetFunc = opts.GetS
@@ -191,7 +187,7 @@ func (s *UnitedSuite) TestWithEnv(c *C) {
 	os.Setenv("TEST_TIMEZONE", "Europe/Berlin")
 	os.Setenv("TEST_LIST", "Test1Env,Test2Env")
 
-	err := Combine(
+	Combine(
 		Mapping{"test:string", "test-string", "TEST_STRING"},
 		Mapping{"test:integer", "test-integer", "TEST_INTEGER"},
 		Mapping{"test:float", "test-float", "TEST_FLOAT"},
@@ -203,8 +199,6 @@ func (s *UnitedSuite) TestWithEnv(c *C) {
 		Mapping{"test:timezone", "test-timezone", "TEST_TIMEZONE"},
 		Mapping{"test:list", "test-list", "TEST_LIST"},
 	)
-
-	c.Assert(err, IsNil)
 
 	c.Assert(GetS("test:string"), Equals, "TestEnv")
 	c.Assert(GetI("test:integer"), Equals, 789)
@@ -230,7 +224,7 @@ func (s *UnitedSuite) TestValidation(c *C) {
 
 	c.Assert(errs, HasLen, 1)
 
-	err := Combine(
+	Combine(
 		Mapping{"test:string", "test-string", "TEST_STRING"},
 		Mapping{"test:integer", "test-integer", "TEST_INTEGER"},
 		Mapping{"test:float", "test-float", "TEST_FLOAT"},
@@ -242,8 +236,6 @@ func (s *UnitedSuite) TestValidation(c *C) {
 		Mapping{"test:timezone", "test-timezone", "TEST_TIMEZONE"},
 		Mapping{"test:list", "test-list", "TEST_LIST"},
 	)
-
-	c.Assert(err, IsNil)
 
 	errs = Validate([]*knf.Validator{
 		{"test:string", knfv.Empty, nil},