+++ knf/united/united.go
@@ -20,7 +20,7 @@ import (
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 // Config is extended configuration
-type Config struct {
+type config struct {
 	mappings map[string]Mapping
 }
 
@@ -38,7 +38,18 @@ type DurationMod = knf.DurationMod
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-var global *Config
+var (
+	Millisecond = knf.Millisecond
+	Second      = knf.Second
+	Minute      = knf.Minute
+	Hour        = knf.Hour
+	Day         = knf.Day
+	Week        = knf.Week
+)
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+var global *config
 
 var optionHasFunc = options.Has
 var optionGetFunc = options.GetS
@@ -47,8 +58,8 @@ var optionGetFunc = options.GetS
 
 // Combine applies mappings to combine knf properties, options, and environment
 // variables
-func Combine(mappings ...Mapping) error {
-	config := &Config{
+func Combine(mappings ...Mapping) {
+	config := &config{
 		mappings: make(map[string]Mapping),
 	}
 
@@ -57,133 +68,83 @@ func Combine(mappings ...Mapping) error {
 	}
 
 	global = config
-
-	return nil
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 // GetS returns configuration value as string
 func GetS(name string, defvals ...string) string {
-	if global == nil {
-		return knf.GetS(name, defvals...)
-	}
-
-	val := global.getProp(name)
-
-	if val == "" && len(defvals) != 0 {
-		return defvals[0]
-	}
-
-	return val
+	return global.GetS(name, defvals...)
 }
 
 // GetI returns configuration value as int
 func GetI(name string, defvals ...int) int {
-	if global == nil {
-		return knf.GetI(name, defvals...)
-	}
-
-	return value.ParseInt(global.getProp(name), defvals...)
+	return global.GetI(name, defvals...)
 }
 
 // GetI64 returns configuration value as int64
 func GetI64(name string, defvals ...int64) int64 {
-	if global == nil {
-		return knf.GetI64(name, defvals...)
-	}
-
-	return value.ParseInt64(global.getProp(name), defvals...)
+	return global.GetI64(name, defvals...)
 }
 
 // GetU returns configuration value as uint
 func GetU(name string, defvals ...uint) uint {
-	if global == nil {
-		return knf.GetU(name, defvals...)
-	}
-
-	return value.ParseUint(global.getProp(name), defvals...)
+	return global.GetU(name, defvals...)
 }
 
 // GetU64 returns configuration value as uint64
 func GetU64(name string, defvals ...uint64) uint64 {
-	if global == nil {
-		return knf.GetU64(name, defvals...)
-	}
-
-	return value.ParseUint64(global.getProp(name), defvals...)
+	return global.GetU64(name, defvals...)
 }
 
 // GetF returns configuration value as floating number
 func GetF(name string, defvals ...float64) float64 {
-	if global == nil {
-		return knf.GetF(name, defvals...)
-	}
-
-	return value.ParseFloat(global.getProp(name), defvals...)
+	return global.GetF(name, defvals...)
 }
 
 // GetB returns configuration value as boolean
 func GetB(name string, defvals ...bool) bool {
-	if global == nil {
-		return knf.GetB(name, defvals...)
-	}
-
-	return value.ParseBool(global.getProp(name), defvals...)
+	return global.GetB(name, defvals...)
 }
 
 // GetM returns configuration value as file mode
 func GetM(name string, defvals ...os.FileMode) os.FileMode {
-	if global == nil {
-		return knf.GetM(name, defvals...)
-	}
-
-	return value.ParseMode(global.getProp(name), defvals...)
+	return global.GetM(name, defvals...)
 }
 
 // GetD returns configuration values as duration
 func GetD(name string, mod DurationMod, defvals ...time.Duration) time.Duration {
-	if global == nil {
-		return knf.GetD(name, mod, defvals...)
-	}
-
-	return value.ParseDuration(global.getProp(name), time.Duration(mod), defvals...)
+	return global.GetD(name, mod, defvals...)
 }
 
 // GetTD returns configuration value as time duration
 func GetTD(name string, defvals ...time.Duration) time.Duration {
-	if global == nil {
-		return knf.GetTD(name, defvals...)
-	}
-
-	return value.ParseTimeDuration(global.getProp(name), defvals...)
+	return global.GetTD(name, defvals...)
 }
 
 // GetTS returns configuration timestamp value as time
 func GetTS(name string, defvals ...time.Time) time.Time {
-	if global == nil {
-		return knf.GetTS(name, defvals...)
-	}
-
-	return value.ParseTimestamp(global.getProp(name), defvals...)
+	return global.GetTS(name, defvals...)
 }
 
 // GetTS returns configuration value as timezone
 func GetTZ(name string, defvals ...*time.Location) *time.Location {
-	if global == nil {
-		return knf.GetTZ(name, defvals...)
-	}
-
-	return value.ParseTimezone(global.getProp(name), defvals...)
+	return global.GetTZ(name, defvals...)
 }
 
 // GetL returns configuration value as list
 func GetL(name string, defvals ...[]string) []string {
+	return global.GetL(name, defvals...)
+}
+
+// Validate executes all given validators and
+// returns slice with validation errors
+func Validate(validators []*knf.Validator) []error {
 	if global == nil {
-		return knf.GetL(name, defvals...)
+		return []error{knf.ErrNilConfig}
 	}
 
-	return value.ParseList(global.getProp(name), defvals...)
+	return validate(validators)
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //
@@ -195,7 +156,132 @@ func (m Mapping) IsEmpty() bool {
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-func (c *Config) getProp(name string) string {
+// GetS returns configuration value as string
+func (c *config) GetS(name string, defvals ...string) string {
+	if c == nil {
+		return knf.GetS(name, defvals...)
+	}
+
+	val := c.getProp(name)
+
+	if val == "" && len(defvals) != 0 {
+		return defvals[0]
+	}
+
+	return val
+}
+
+// GetI returns configuration value as int
+func (c *config) GetI(name string, defvals ...int) int {
+	if c == nil {
+		return knf.GetI(name, defvals...)
+	}
+
+	return value.ParseInt(c.getProp(name), defvals...)
+}
+
+// GetI64 returns configuration value as int64
+func (c *config) GetI64(name string, defvals ...int64) int64 {
+	if c == nil {
+		return knf.GetI64(name, defvals...)
+	}
+
+	return value.ParseInt64(c.getProp(name), defvals...)
+}
+
+// GetU returns configuration value as uint
+func (c *config) GetU(name string, defvals ...uint) uint {
+	if c == nil {
+		return knf.GetU(name, defvals...)
+	}
+
+	return value.ParseUint(c.getProp(name), defvals...)
+}
+
+// GetU64 returns configuration value as uint64
+func (c *config) GetU64(name string, defvals ...uint64) uint64 {
+	if c == nil {
+		return knf.GetU64(name, defvals...)
+	}
+
+	return value.ParseUint64(c.getProp(name), defvals...)
+}
+
+// GetF returns configuration value as floating number
+func (c *config) GetF(name string, defvals ...float64) float64 {
+	if c == nil {
+		return knf.GetF(name, defvals...)
+	}
+
+	return value.ParseFloat(c.getProp(name), defvals...)
+}
+
+// GetB returns configuration value as boolean
+func (c *config) GetB(name string, defvals ...bool) bool {
+	if c == nil {
+		return knf.GetB(name, defvals...)
+	}
+
+	return value.ParseBool(c.getProp(name), defvals...)
+}
+
+// GetM returns configuration value as file mode
+func (c *config) GetM(name string, defvals ...os.FileMode) os.FileMode {
+	if c == nil {
+		return knf.GetM(name, defvals...)
+	}
+
+	return value.ParseMode(c.getProp(name), defvals...)
+}
+
+// GetD returns configuration values as duration
+func (c *config) GetD(name string, mod DurationMod, defvals ...time.Duration) time.Duration {
+	if c == nil {
+		return knf.GetD(name, mod, defvals...)
+	}
+
+	return value.ParseDuration(c.getProp(name), time.Duration(mod), defvals...)
+}
+
+// GetTD returns configuration value as time duration
+func (c *config) GetTD(name string, defvals ...time.Duration) time.Duration {
+	if c == nil {
+		return knf.GetTD(name, defvals...)
+	}
+
+	return value.ParseTimeDuration(c.getProp(name), defvals...)
+}
+
+// GetTS returns configuration timestamp value as time
+func (c *config) GetTS(name string, defvals ...time.Time) time.Time {
+	if c == nil {
+		return knf.GetTS(name, defvals...)
+	}
+
+	return value.ParseTimestamp(c.getProp(name), defvals...)
+}
+
+// GetTS returns configuration value as timezone
+func (c *config) GetTZ(name string, defvals ...*time.Location) *time.Location {
+	if c == nil {
+		return knf.GetTZ(name, defvals...)
+	}
+
+	return value.ParseTimezone(c.getProp(name), defvals...)
+}
+
+// GetL returns configuration value as list
+func (c *config) GetL(name string, defvals ...[]string) []string {
+	if c == nil {
+		return knf.GetL(name, defvals...)
+	}
+
+	return value.ParseList(c.getProp(name), defvals...)
+}
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+func (c *config) getProp(name string) string {
 	m := c.mappings[name]
 
 	if m.IsEmpty() {
@@ -210,4 +296,18 @@ func (c *Config) getProp(name string) string {
 	default:
 		return knf.GetS(name)
 	}
+}
+
+func validate(validators []*knf.Validator) []error {
+	var result []error
+
+	for _, v := range validators {
+		err := v.Func(global, v.Property, v.Value)
+
+		if err != nil {
+			result = append(result, err)
+		}
+	}
+
+	return result
 }